Move fronts-specific code to another component to prevent excessive re-rendering
